Generalize PasswordRule enum into a simple wrapper. Factor out rules-based and entropy-based PasswordValidator implementation hierarchies.
